requirements.txt
Flask==1.0.2
Flask-WTF==0.14.2
Keras==2.2.4
Keras-Applications==1.0.6
numpy==1.15.4
Pillow==5.4.1
tensorflow==1.12.0
WTForms==2.2.1